price_final should also be updated when changing price for sale
git-svn-id: http://62.90.102.41/svn/NiceHouse/trunk@1536 1431c75d-6ac1-4f6d-bc36-6b0d284ea005

+++ Management/models.py
@@ -2042,7 +2042,7 @@ class NHSale(models.Model):
     num = models.IntegerField(ugettext('sale_num'))
     address = models.CharField(ugettext('address'), max_length=50)
     hood = models.CharField(ugettext('hood'), max_length=50)
-    rooms = models.PositiveSmallIntegerField(ugettext('rooms'))
+    rooms = models.FloatField(ugettext('rooms'))
     floor = models.PositiveSmallIntegerField(ugettext('floor'))
     type = models.ForeignKey('HouseType', verbose_name = ugettext('house_type'))
